Edit userinfo,signout
